Delete EncryptedDM module, fold into Client.Send.encrypted_dm/4; nits, docs, formatting
